Camera Source Select Function 6
This finally works! I think! I've rewritten the functions so that the camera sources are now accessed by a select element. When the element is changed the getUserMedia constraint is modified to take the id of the selected camera.

The app is launched when you switch camera sources. Pressing the start app button appends the video and its container to the main container and then launched fullscreen.